+++ packages/backend/src/apps/stripe/auth/verify-credentials.ts
@@ -1,13 +1,9 @@
 import { IGlobalVariable } from '@automatisch/types';
 
 const verifyCredentials = async ($: IGlobalVariable) => {
-  try {
-    await $.http.get(
-        `/v1/events`,
-    );
-  } catch (e) {
-    throw new Error('Invalid secret key')
-  }
+  await $.http.get(
+      `/v1/events`,
+  );
   await $.auth.set({
     screenName: $.auth.data?.displayName,
   });
