+++ packages/backend/src/apps/stripe/common/add-auth-header.ts
@@ -1,8 +1,8 @@
 import {TBeforeRequest} from "@automatisch/types";
 
 const addAuthHeader: TBeforeRequest = ($, requestConfig) => {
-    requestConfig.headers['Authorization'] = `Bearer ${$.auth.data?.secretKey}`
-    return requestConfig
+  requestConfig.headers['Authorization'] = `Bearer ${$.auth.data?.secretKey}`
+  return requestConfig
 }
 
 export default addAuthHeader;
