+++ packages/backend/src/apps/stripe/triggers/balance-transaction/index.ts
@@ -1,12 +1,12 @@
 import defineTrigger from "../../../../helpers/define-trigger";
-import getBalanceTransactions from "../../common/get-balance-transactions";
+import getBalanceTransactions from "./get-balance-transactions";
 
 export default defineTrigger({
-    name: 'Balance Transaction',
-    key: 'balanceTransaction',
-    description: 'Triggers when a new transaction is processed (refund, payout, adjustment, ...)',
-    pollInterval: 15,
-    async run($) {
-        await getBalanceTransactions($)
-    }
+  name: 'New Balance Transactions',
+  key: 'balanceTransaction',
+  description: 'Triggers when a new transaction is processed (refund, payout, adjustment, ...)',
+  pollInterval: 15,
+  async run($) {
+    await getBalanceTransactions($)
+  }
 })
