+++ packages/backend/src/apps/stripe/triggers/payouts/index.ts
@@ -1,12 +1,12 @@
 import defineTrigger from "../../../../helpers/define-trigger";
-import getPayouts from "../../common/get-payouts";
+import getPayouts from "./get-payouts";
 
 export default defineTrigger({
-    name: 'Payout',
-    key: 'payout',
-    description: 'Triggers when a payout (Stripe <-> Bank account) has been updated',
-    pollInterval: 15,
-    async run($) {
-        await getPayouts($)
-    }
+  name: 'New Payouts',
+  key: 'payout',
+  description: 'Triggers when a payout (Stripe <-> Bank account) has been updated',
+  pollInterval: 15,
+  async run($) {
+    await getPayouts($)
+  }
 })
