+++ packages/docs/pages/apps/stripe/connection.md
@@ -9,6 +9,6 @@ You are free to use the **Testing secret key** instead of the productive secret 
 :::
 
 1. Go to the [Stripe Dashboard > Developer > API keys](https://dashboard.stripe.com/apikeys)
-2. Click on **Reveal live key** in the table row **Secret key** and copy the now shown secret key 3
+2. Click on **Reveal live key** in the table row **Secret key** and copy the now shown secret key
 3. Paste the **Secret key** in the named field in Automatisch and assign a display name for the connection.
 4. Congrats! You can start using the new Stripe connection!
